requirements.txt
# === Core Dependencies ===
python-dotenv>=1.0.0

# === Web Scraping ===
playwright>=1.45.0
beautifulsoup4>=4.12.0

# === PDF Processing ===
pypdf>=4.0.0

# === AI/ML Stack ===
# LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0

# Embeddings
sentence-transformers>=2.7.0

# Vector Stores
faiss-cpu>=1.8.0
chromadb>=0.5.0

# OpenAI Compatible (DeepSeek)
openai>=1.50.0

# === Interface ===
streamlit>=1.38.0

# === Automation ===
schedule>=1.2.0

# === Optional: OCR pour PDFs scannés ===
# Décommentez si nécessaire :
# pytesseract>=0.3.10
# pdf2image>=1.16.0
# Pillow>=10.0.0

# === Optional: Embeddings Français Optimisés ===
# Décommentez pour améliorer la qualité en français :
# transformers>=4.40.0
# torch>=2.0.0

# === Development ===
# pytest>=8.0.0
# black>=24.0.0
